Add dockerfile for building arduino project
ChatGPT was used to generate some parts of the file. It was very good at finding all required packages to be able to run the arduino-cli, which was very useful as it would have taken considetable effort for me to research that on my own. It also helped a bit by figuring out how to install the arduino-cli as it knows the commands for unzipping and different curl flags. Most of the code it generated did, however, need considerable modification